requirements.txt
# Jarvis AI Assistant Dependencies
google-generativeai>=0.3.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
PyAudio>=0.2.14
